Extract ClientFlowController and ServerFlowController headers
