+++ src/core/socket/stream/ClientFlowController.h
@@ -1,0 +1,128 @@
+/*
+ * SNode.C - A Slim Toolkit for Network Communication
+ * Copyright (C) Volker Christian <me@vchrist.at>
+ *               2020, 2021, 2022, 2023, 2024, 2025, 2026
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published
+ * by the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+/*
+ * MIT License
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+
+#ifndef CORE_SOCKET_STREAM_CLIENTFLOWCONTROLLER_H
+#define CORE_SOCKET_STREAM_CLIENTFLOWCONTROLLER_H
+
+#include "core/EventReceiver.h"
+#include "core/socket/stream/FlowController.h"
+#include "core/timer/Timer.h"
+
+#ifndef DOXYGEN_SHOULD_SKIP_THIS
+
+#include <functional>
+#include <memory>
+
+#endif /* DOXYGEN_SHOULD_SKIP_THIS */
+
+namespace core::socket::stream {
+
+    class ClientFlowController final : public core::socket::stream::FlowController<ClientFlowController> {
+    public:
+        explicit ClientFlowController(net::config::ConfigInstance* configInstance)
+            : core::socket::stream::FlowController<ClientFlowController>(configInstance)
+            , onFlowReconnectCallback([](ClientFlowController*) {
+            }) {
+        }
+
+        void stopReconnect() {
+            reconnectEnabled = false;
+            cancelReconnectTimer();
+        }
+
+        bool isReconnectEnabled() const {
+            return reconnectEnabled;
+        }
+
+        ClientFlowController* onFlowReconnect(const std::function<void(ClientFlowController*)>& callback) {
+            const std::function<void(ClientFlowController*)> oldCallback = onFlowReconnectCallback;
+            onFlowReconnectCallback = [oldCallback, callback](ClientFlowController* flowController) {
+                oldCallback(flowController);
+                callback(flowController);
+            };
+
+            return this;
+        }
+
+        void reportFlowReconnect() {
+            onFlowReconnectCallback(this);
+        }
+
+        void observeConnectEventReceiver(core::eventreceiver::ConnectEventReceiver* connectEventReceiver) {
+            if (connectEventReceiver != nullptr && connectEventReceiver->isEnabled()) {
+                this->connectEventReceiver = connectEventReceiver;
+            } else {
+                this->connectEventReceiver = nullptr;
+            }
+        }
+
+        void armReconnectTimer(double timeoutSeconds, const std::function<void()>& dispatcher) {
+            if (reconnectEnabled) {
+                reconnectTimer = std::make_unique<core::timer::Timer>(core::timer::Timer::singleshotTimer(dispatcher, timeoutSeconds));
+            }
+        }
+
+    private:
+        void terminateAsyncSubFlow() override {
+            stopReconnect();
+            this->stopRetry();
+
+            if (connectEventReceiver != nullptr) {
+                connectEventReceiver->stopConnect();
+            }
+        }
+
+        void cancelReconnectTimer() {
+            if (reconnectTimer) {
+                reconnectTimer->cancel();
+                reconnectTimer.reset();
+            }
+        }
+
+        bool reconnectEnabled{true};
+        core::eventreceiver::ConnectEventReceiver* connectEventReceiver{nullptr};
+        std::unique_ptr<core::timer::Timer> reconnectTimer;
+        std::function<void(ClientFlowController*)> onFlowReconnectCallback;
+    };
+
+} // namespace core::socket::stream
+
+#endif // CORE_SOCKET_STREAM_CLIENTFLOWCONTROLLER_H
